+++ Classes/Contact.h
@@ -28,6 +28,7 @@
 - (instancetype)initWithPerson:(ABRecordRef)person;
 - (instancetype)initWithFriend:(LinphoneFriend *) friend;
 
+- (void)setAvatar:(UIImage *)avatar;
 - (BOOL)setSipAddress:(NSString *)sip atIndex:(NSInteger)index;
 - (BOOL)setEmail:(NSString *)email atIndex:(NSInteger)index;
 - (BOOL)setPhoneNumber:(NSString *)phone atIndex:(NSInteger)index;
@@ -39,4 +40,7 @@
 - (BOOL)removeSipAddressAtIndex:(NSInteger)index;
 - (BOOL)removePhoneNumberAtIndex:(NSInteger)index;
 - (BOOL)removeEmailAtIndex:(NSInteger)index;
+
+- (int)remove;
+- (BOOL)save;
 @end